Update to 2.0
Update the demo to v2.0, this also updates the engine to 4.4.1, Terrain3D has also been updated.
This update also includes a lot of cleaning in the files.

This version has improved tomemapping compared to 2.0 on itch.io, as this version uses AgX

I hope this can be useful for some!